requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
billiard==4.1.0
black==23.9.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.0
cron-descriptor==1.4.0
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-templated-mail==1.1.1
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
executing==1.2.0
flake8==6.1.0
idna==3.4
inflection==0.5.1
ipython==8.15.0
isort==5.12.0
jedi==0.19.0
kombu==5.3.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
